batch_emailer/reference_answer/batch-emailer/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
jinja2==3.1.4
